+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### Fixes
 
+- Resources could be changed by using the copy-of constructor of the builders, though they are immutable.
 - A `400 BAD REQUEST` response now creates a `BadRequestException` instead of a `ConflictException` 
 
 ## 1.8 - 2015-12-12